+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix Container::attach output when TTY is enabled on container
 - Add `Image::build_par` that uses multithreaded compression algorithm for creating context directory archive. This method is behind `par-compression` feature flag.
 - Add `ImageListOptsBuilder::shared_size` parameter
+- Add `SystemDataUsageOpts` parameter to `Docker::data_usage`
 
 # 0.12.0
 - Fix some integer fields that could be negative but previously were a usize like `ImageSummary::containers`